% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{Add_Spatial_Effort}
\alias{Add_Spatial_Effort}
\title{Add Fishing Effort Spatial Distribution to MOM}
\usage{
Add_Spatial_Effort(MOM, Spatial_Effort)
}
\arguments{
\item{MOM}{An object of class \code{MOM}}

\item{Spatial_Effort}{A data.frame with Spatial Effort by Area and fleet.
If multiple stocks, \code{Spatial_Effort} should be a list length \code{nstocks}}
}
\value{
An updated MOM object
}
\description{
Add Fishing Effort Spatial Distribution to MOM
}
