% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Fleets.R
\name{AddSeasonal_RS}
\alias{AddSeasonal_RS}
\title{Genarate OM with On- and Off-Season Fleets - Red Snapper}
\usage{
AddSeasonal_RS(
  MOM,
  relF,
  On.Fleet = 1,
  Off.Fleet = 4,
  Fleet = "Commercial Handline"
)
}
\arguments{
\item{MOM}{An object of class \code{MOM}}

\item{relF}{A dataframe with the relative F for On- and Off-Season for each year}

\item{On.Fleet}{Index for On-Season Fleet}

\item{Off.Fleet}{Index for Off-Season Fleet}

\item{Fleet}{Name of the Fleet}
}
\value{
An object of class \code{MOM}
}
\description{
Genarate OM with On- and Off-Season Fleets - Red Snapper
}
