% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{Aggregate_Fleets}
\alias{Aggregate_Fleets}
\title{Compare the F trends from BAM and openMSE}
\usage{
Aggregate_Fleets(MOM, fleet_df, discard_mortality)
}
\arguments{
\item{MOM}{An object of class \code{MOM}}

\item{fleet_df}{A dataframe. See 'Operating Models' article}

\item{discard_mortality}{A dataframe. See 'Operating Models' article}
}
\value{
An updated MOM
}
\description{
Compare the F trends from BAM and openMSE
}
