% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference_Points.R
\name{Calc_RS_Ref_Points}
\alias{Calc_RS_Ref_Points}
\title{Calculate Reference Points for Red Snapper}
\usage{
Calc_RS_Ref_Points(x, y, multiHist)
}
\arguments{
\item{multiHist}{}
}
\description{
Calculate Reference Points for Red Snapper
}
