% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Fleets.R
\name{calc_seasons_RS_Comm}
\alias{calc_seasons_RS_Comm}
\alias{calc_seasons_RS_HB}
\title{Calculate removals by season for Red Snapper Commercial handline}
\usage{
calc_seasons_RS_Comm(year, logbook, season_df)

calc_seasons_RS_HB(year, logbook, season_df)
}
\arguments{
\item{year}{numeric year}

\item{logbook}{commercial logbook dataframe}

\item{season_df}{season dataframe}
}
\value{
data.frame
}
\description{
Calculate removals by season for Red Snapper Commercial handline
}
\section{Functions}{
\itemize{
\item \code{calc_seasons_RS_HB()}: Calculate season for Red Snapper Headboat

}}
