% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference_Points.R
\name{Calculate_Ref_Points}
\alias{Calculate_Ref_Points}
\title{Calculate Biological Reference Points}
\usage{
Calculate_Ref_Points(multiHist)
}
\arguments{
\item{multiHist}{}
}
\description{
Calculate Biological Reference Points
}
