% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_OMs.R
\name{Combine_OMs}
\alias{Combine_OMs}
\title{Combine Multiple Single-Stock MOM objects into one Multi-Stock MOM}
\usage{
Combine_OMs(OMlist, Name = NULL, truncsd = 2)
}
\arguments{
\item{OMlist}{A list of \code{MOM} objects}

\item{Name}{Name of the output MOM object}

\item{truncsd}{Numeric value for number of standard deviations to truncate
correlated recruitment deviations. Set to NULL to ignore.}
}
\value{
An object of class \code{\link[=MOM]{MOM()}}
}
\description{
Combine Multiple Single-Stock MOM objects into one Multi-Stock MOM
}
