% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Fleets.R
\name{combineFleets}
\alias{combineFleets}
\title{Combine Discard and Landing Fleets in an MOM object}
\usage{
combineFleets(MOM, stock = 1, fleets = list(c(1, 4), c(2, 5), c(3, 6)))
}
\arguments{
\item{MOM}{An object of class \code{MOM}}

\item{stock}{The stock number (currently only works for 1 stock)}

\item{fleets}{List with each element numeric vector length 2 to indicate landing and discard fleet. Length 1 means only a landing fleet}
}
\value{
An object of class \code{MOM} with \code{length(fleets)} fleets
}
\description{
Combine Discard and Landing Fleets in an MOM object
}
