% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{Compare_Biomass}
\alias{Compare_Biomass}
\title{Compare the biomass trends from BAM and openMSE}
\usage{
Compare_Biomass(Hist, rdat)
}
\arguments{
\item{Hist}{An object of class \code{multiHist}}

\item{rdat}{A list object from bamExtras}
}
\value{
a ggplot object
}
\description{
Compare the biomass trends from BAM and openMSE
}
