% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{Compare_Catch}
\alias{Compare_Catch}
\title{Compare the landings & discards from BAM and openMSE}
\usage{
Compare_Catch(Hist, rdat, divide = FALSE)
}
\arguments{
\item{Hist}{An object of class \code{multiHist}}

\item{rdat}{A list object from bamExtras}
}
\value{
a ggplot object
}
\description{
Compare the landings & discards from BAM and openMSE
}
