% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{Compare_F}
\alias{Compare_F}
\title{Compare the F trends from BAM and openMSE}
\usage{
Compare_F(Hist, rdat)
}
\arguments{
\item{Hist}{An object of class \code{multiHist}}

\item{rdat}{A list object from bamExtras}
}
\value{
a ggplot object
}
\description{
Compare the F trends from BAM and openMSE
}
