% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{fixedC_mean3}
\alias{fixedC_mean3}
\title{Example fixed catch MP (removals)}
\usage{
fixedC_mean3(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed catch MP (removals)
}
