% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{fixedF_mean3}
\alias{fixedF_mean3}
\title{Example fixed F MP}
\usage{
fixedF_mean3(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP
}
