% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{fixedF_mean3_MLL}
\alias{fixedF_mean3_MLL}
\title{Example fixed F MP with MLL}
\usage{
fixedF_mean3_MLL(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP with MLL
}
