% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_MMP.R
\name{Fleet_MMP}
\alias{Fleet_MMP}
\title{Generic MMP for SAMFC}
\usage{
Fleet_MMP(x, DataList, Fleet_Management, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A nested list of \code{Data} objects}

\item{Fleet_Management}{data.frame with management actions (F, Catch, and/or MLL) for this management cycle}

\item{...}{}
}
\value{
A list of \code{Rec} objects
}
\description{
Generic MMP for SAMFC
}
