% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{Ftarget}
\alias{Ftarget}
\title{Example fixed F MP that sets the F for each stock to the MFMT}
\usage{
Ftarget(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP that sets the F for each stock to the MFMT
}
