% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Projection_Rec_Devs.R
\name{Generate_Future_Rec_Devs}
\alias{Generate_Future_Rec_Devs}
\title{Update Recruitment Deviations in the projection years}
\usage{
Generate_Future_Rec_Devs(MOM, truncsd = 2)
}
\arguments{
\item{MOM}{An MOM object with cpars populated}

\item{truncsd}{Integer. The number of standard deviations to truncate}
}
\value{
An updated MOM
}
\description{
Sampling from a multivariate truncated normal distribution
}
