% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{get_MLL}
\alias{get_MLL}
\title{Return the minimum legal length (MLL) for a given stock}
\usage{
get_MLL(Name)
}
\arguments{
\item{Name}{The common name of the stock}
}
\value{
numeric value. MLL in mm
}
\description{
Return the minimum legal length (MLL) for a given stock
}
