% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{get_rebuild_year}
\alias{get_rebuild_year}
\title{Get the rebuilding target year}
\usage{
get_rebuild_year(stock_code)
}
\arguments{
\item{stock_code}{either 'RS', 'GG', or 'BS'}
}
\description{
Get the rebuilding target year
}
