% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{Landings_Removals}
\alias{Landings_Removals}
\title{Mean ratio of landings to total removals}
\usage{
Landings_Removals(MMSE)
}
\arguments{
\item{MMSE}{}
}
\description{
Mean ratio of landings to total removals
}
