% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{lower_rec_effort}
\alias{lower_rec_effort}
\title{Reduce Gen Rec Landings & Discards and run BAM model}
\usage{
lower_rec_effort(CommonName, reduce_frac, bam_name = NULL)
}
\arguments{
\item{CommonName}{Common name for a stock in \code{bamExtras}}

\item{reduce_frac}{Fractions to reduce the landings & discards}
}
\value{
the output of \code{bamExtras::run_bam}
}
\description{
Reduce Gen Rec Landings & Discards and run BAM model
}
