% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OMs.R
\name{modify_adult_M}
\alias{modify_adult_M}
\title{Modify M and run BAM model}
\usage{
modify_adult_M(CommonName, new_M, bam_name = NULL)
}
\arguments{
\item{CommonName}{Common name for a stock in \code{bamExtras}}

\item{new_M}{New value for adult M}
}
\value{
the output of \code{bamExtras::run_bam}
}
\description{
Modify M and run BAM model
}
