% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{P_rebuild}
\alias{P_rebuild}
\title{Probability of rebuilding to target level by end of projection period}
\usage{
P_rebuild(
  MMSE = NULL,
  SSBtarg = list(`Red Snapper` = 680555, `Gag Grouper` = 1772)
)
}
\arguments{
\item{MMSE}{}
}
\description{
Probability of rebuilding to target level by end of projection period
}
