% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_Catch}
\alias{plot_Catch}
\title{Plot Landings and Discards from Projection Period}
\usage{
plot_Catch(MMSE, colors = c("darkblue", "darkred"), inc.hist = FALSE)
}
\arguments{
\item{inc.hist}{Include historical period?}
}
\description{
Plot Landings and Discards from Projection Period
}
