% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_Catch_Discards}
\alias{plot_Catch_Discards}
\title{Plot Landings and Discards}
\usage{
plot_Catch_Discards(multiHist)
}
\arguments{
\item{multiHist}{}
}
\value{
A ggplot object
}
\description{
Plot Landings and Discards
}
