% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_Fmort}
\alias{plot_Fmort}
\title{Plot Total Fishing Mortality}
\usage{
plot_Fmort(MMSE, inc.hist = TRUE)
}
\arguments{
\item{inc.hist}{Include historical period?}
}
\description{
Plot Total Fishing Mortality
}
