% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_SB}
\alias{plot_SB}
\title{Plot Spawning Stock and Reference Pints}
\usage{
plot_SB(MMSE, inc.hist = T)
}
\arguments{
\item{inc.hist}{Include historical period?}
}
\description{
Plot Spawning Stock and Reference Pints
}
