% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_SB_hist}
\alias{plot_SB_hist}
\title{Plot Landings and Discards}
\usage{
plot_SB_hist(
  SSBhist,
  Ref_Points,
  inc_ref_point = TRUE,
  rel_to = NA,
  ymax = list()
)
}
\arguments{
\item{multiHist}{}
}
\value{
A ggplot object
}
\description{
Plot Landings and Discards
}
