% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{Process_Results}
\alias{Process_Results}
\title{Create data.frames of MSE results and save to disk}
\usage{
Process_Results(om = "BaseCase", MSE_dir = "MSE_Objects")
}
\arguments{
\item{om}{OM name}

\item{MSE_dir}{}
}
\value{
Nothing. Saves data.frames to 'Results_Objects'
}
\description{
Create data.frames of MSE results and save to disk
}
