% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{StatusQuo}
\alias{StatusQuo}
\title{Example fixed F MP that sets F for each fleet to the mean from the last 3 historical years}
\usage{
StatusQuo(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP that sets F for each fleet to the mean from the last 3 historical years
}
