% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_OM_fleets.R
\name{structure_OM}
\alias{structure_OM}
\title{Create Seasonal Fleet Structured OM from BAM Assessment}
\usage{
structure_OM(
  rdat,
  stock_name,
  nsim,
  pyears,
  CAL_bins,
  Fleet_Structure,
  seasonal_F_List
)
}
\arguments{
\item{rdat}{An \code{rdat} object from the \code{BAMextras} package}

\item{stock_name}{Name of the stock object}

\item{nsim}{Number of simulations in the MSE}

\item{pyears}{Number of projection years in the MSE}

\item{CAL_bins}{The Catch-at-Length bins used to generate CAL data. Must be the
same across all stocks and fleets in a multi-species OM}

\item{Fleet_Structure}{A data.frame with the name of the fleets, and index
values for the position of the Landing and Discard for each fleet}

\item{seasonal_F_List}{A list of n fleets long, with each element containing
a dataframe with the relative F for the On-Season and Off-Season fleets, the discard
mortality, the Discard rate, and the name of the fleet}
}
\value{
an object of class \code{MOM} with fleets structure into On and Off Seasons
}
\description{
Currently only tested for Red Snapper and Gag
}
\details{
Should work for other assessments, provided they have a similar
fleet structure to RS and Gag, i.e., Commercial Handline, Recreational Headboat,
and General Recreational Landing and Discards fleets
}
