% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{TradeOff}
\alias{TradeOff}
\title{Trade-Off Plot of Two Performance Metrics}
\usage{
TradeOff(MMSE, PMs = c("P_MSST", "P_MFMT"))
}
\arguments{
\item{PMs}{}
}
\description{
Trade-Off Plot of Two Performance Metrics
}
