% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{Write_MPs}
\alias{Write_MPs}
\alias{Source_MPs}
\title{Write MP functions}
\usage{
Write_MPs(MPList, Rec_Reduction, dir = "MP_functions")

Source_MPs(dir = "MP_functions")
}
\arguments{
\item{MPList}{A character string with names of the MPs to create}

\item{Rec_Reduction}{A numeric vector with effort reductions for General
Recreational fleet}

\item{dir}{Directory to write the MPs to}
}
\value{
A list of MPs names
}
\description{
Write MP functions
}
\section{Functions}{
\itemize{
\item \code{Source_MPs()}: Source the MPs

}}
