% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_OMs.R
\name{Combine_OMs}
\alias{Combine_OMs}
\title{Combine Multiple Single-Stock MOM objects into one Multi-Stock MOM}
\usage{
Combine_OMs(OMList, Name = NULL, fill_yrs = 3)
}
\arguments{
\item{OMList}{A list of \code{MOM} objects}

\item{Name}{Name of the output MOM object}

\item{fill_yrs}{Number of years to calculate mean values to fill fishing effort (\code{cpars$Find}),
for any OMs that have a terminal year before the maximum terminal year of all OMs}
}
\value{
An object of class \code{\link[=MOM]{MOM()}}
}
\description{
Combine Multiple Single-Stock MOM objects into one Multi-Stock MOM
}
