% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{Ftarget}
\alias{Ftarget}
\alias{Ftarget_IncRecEff}
\title{Example fixed F MP that sets the F for each stock to the MFMT}
\usage{
Ftarget(x, DataList, incE = list(), MLL = list(), ...)

Ftarget_IncRecEff(x, DataList, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP that sets the F for each stock to the MFMT
}
\section{Functions}{
\itemize{
\item \code{Ftarget_IncRecEff()}: Ftarget with General Recreational Effort increasing by 2\% per year

}}
