% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Projection_Rec_Devs.R
\name{Generate_Correlated_Rec_Devs}
\alias{Generate_Correlated_Rec_Devs}
\alias{Plot_Correlated_Rec_Devs}
\title{Update Recruitment Deviations in the projection years}
\usage{
Generate_Correlated_Rec_Devs(OMList, truncsd = 2)

Plot_Correlated_Rec_Devs(OMList, ncol = 3)
}
\arguments{
\item{OMList}{A list of MOM objects with cpars populated}

\item{truncsd}{Integer. The number of standard deviations to truncate recruitment deviations}

\item{ncol}{Number of columns for the plot}
}
\value{
An updated A list of MOM objects with rec devs updated in cpars
}
\description{
Sampling from a multivariate truncated normal distribution
}
\section{Functions}{
\itemize{
\item \code{Plot_Correlated_Rec_Devs()}: Scatter plot with marginal histograms

}}
