% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{MPs}
\alias{MPs}
\alias{Create_Rec_List}
\alias{Get_Year_Info}
\alias{SQ}
\alias{Adjust_Effort}
\alias{SQ_FR}
\alias{SQ_MLL}
\alias{SQ_NS}
\alias{SQ_OS}
\alias{SQ_FR_MLL}
\alias{SQ_FR_MLL_NS}
\alias{SQ_FR_MLL_OS}
\title{Management Procedures}
\usage{
Create_Rec_List(DataList)

Get_Year_Info(DataList)

SQ(x, DataList, ...)

Adjust_Effort(RecList, Effort_Mod = 0, ...)

SQ_FR(x, DataList, First_Management_Year = 2025, ...)

SQ_MLL(x, DataList, First_Management_Year = 2025, ...)

SQ_NS(x, DataList, First_Management_Year = 2025, ...)

SQ_OS(x, DataList, First_Management_Year = 2025, ...)

SQ_FR_MLL(x, DataList, First_Management_Year = 2025, ...)

SQ_FR_MLL_NS(x, DataList, ...)

SQ_FR_MLL_OS(x, DataList, ...)
}
\arguments{
\item{DataList}{A nested list of \code{Data} objects}

\item{x}{Simulation number}

\item{...}{Additional arguments}

\item{RecList}{A list of \code{Rec} objects returned by an \code{MMP}}

\item{Effort_Mod}{Effort modifier. Effort is adjusted as
\code{Effort_Rec_Fleet <- Effort_Rec_Fleet * (1-Effort_Mod)}}

\item{First_Management_Year}{Year the management procedures are first implemented}
}
\description{
Management Procedures
}
\section{Functions}{
\itemize{
\item \code{Create_Rec_List()}: Generate a list of \code{Rec} objects for an \code{MMP}

\item \code{Get_Year_Info()}: Return information on current year from a \code{DataList}

\item \code{SQ()}: Status Quo. Fishing effort for each fleet is fixed in the projection
years at the geometric mean of the three last historical years

\item \code{Adjust_Effort()}: Adjust fishing effort for General Recreational Fleet

\item \code{SQ_FR()}: Status Quo with Full Retention. Modify retention curve with an asymptote of 1

\item \code{SQ_MLL()}: Status Quo with a stock specific size limit.

\item \code{SQ_NS()}: Status Quo with offshore areas closed to fishing

\item \code{SQ_OS()}: Status Quo with nearshore areas closed to fishing

\item \code{SQ_FR_MLL()}: Status Quo with full retention and minimum size limit

\item \code{SQ_FR_MLL_NS()}: Status Quo with full retention, minimum size limit, and offshore closed

\item \code{SQ_FR_MLL_OS()}: Status Quo with full retention, minimum size limit, and nearshore closed

}}
