% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{Plot_Prob_Rebuild}
\alias{Plot_Prob_Rebuild}
\alias{PM_plots}
\alias{SSB_plot}
\title{PM Plotting Functions}
\usage{
Plot_Prob_Rebuild(
  om = "BaseCase",
  results_dir = "Results_Objects",
  addtheme = NULL
)

PM_plots(
  om = "BaseCase",
  stock_code = "RS",
  results_dir = "Results_Objects",
  textsize = 2
)

SSB_plot(om = "BaseCase", stock_code = "RS", results_dir = "Results_Objects")
}
\arguments{
\item{results_dir}{}
}
\description{
PM Plotting Functions
}
\section{Functions}{
\itemize{
\item \code{PM_plots()}: Plots of other performance metrics

\item \code{SSB_plot()}: SSB/MSST time-series plot

}}
