% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{StatusQuo}
\alias{StatusQuo}
\alias{StatusQuo_IncRecEff}
\alias{RecEff20}
\alias{RecEff20_IncRecEff}
\alias{RecEff40}
\alias{RecEff40_IncRecEff}
\alias{MLL20_25}
\alias{MLL20_25_IncRecEff}
\alias{MLL25_25}
\alias{MLL25_25_IncRecEff}
\title{Example fixed F MP that sets F for each fleet to the mean from the last 3 historical years}
\usage{
StatusQuo(x, DataList, incE = list(), MLL = list(), Eff_multi = list(), ...)

StatusQuo_IncRecEff(
  x,
  DataList,
  incE = list(`General Recreational` = 0.02),
  ...
)

RecEff20(
  x,
  DataList,
  incE = list(),
  MLL = list(),
  Eff_multi = list(`General Recreational` = 0.8),
  ...
)

RecEff20_IncRecEff(
  x,
  DataList,
  incE = list(`General Recreational` = 0.02),
  MLL = list(),
  Eff_multi = list(`General Recreational` = 0.8),
  ...
)

RecEff40(
  x,
  DataList,
  incE = list(),
  MLL = list(),
  Eff_multi = list(`General Recreational` = 0.6),
  ...
)

RecEff40_IncRecEff(
  x,
  DataList,
  incE = list(`General Recreational` = 0.02),
  MLL = list(),
  Eff_multi = list(`General Recreational` = 0.6),
  ...
)

MLL20_25(
  x,
  DataList,
  incE = list(),
  MLL = list(`Red Snapper` = 20, `Gag Grouper` = 25),
  ...
)

MLL20_25_IncRecEff(
  x,
  DataList,
  incE = list(`General Recreational` = 0.02),
  MLL = list(`Red Snapper` = 20, `Gag Grouper` = 25),
  ...
)

MLL25_25(
  x,
  DataList,
  incE = list(),
  MLL = list(`Red Snapper` = 25, `Gag Grouper` = 25),
  ...
)

MLL25_25_IncRecEff(
  x,
  DataList,
  incE = list(`General Recreational` = 0.02),
  MLL = list(`Red Snapper` = 25, `Gag Grouper` = 25),
  ...
)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list of \code{Data} objects}

\item{...}{Additional arguments. Not used}
}
\value{
A list of \code{Rec} objects
}
\description{
Example fixed F MP that sets F for each fleet to the mean from the last 3 historical years
}
\section{Functions}{
\itemize{
\item \code{StatusQuo_IncRecEff()}: Status Quo MP with General Recreational Effort increasing by 2\% per year

\item \code{RecEff20()}: StatusQuo with Effort for General Recreational Fleet reduce by 20\% for both stocks

\item \code{RecEff20_IncRecEff()}: RecEff20 with General Recreational Effort increasing by 2\% per year

\item \code{RecEff40()}: StatusQuo with Effort for General Recreational Fleet reduce by 40\% for both stocks

\item \code{RecEff40_IncRecEff()}: RecEff40 with General Recreational Effort increasing by 2\% per year

\item \code{MLL20_25()}: StatusQuo with Minimum Legal Length for both stocks (20 inch for Red Snapper and 25 inch for Gag)

\item \code{MLL20_25_IncRecEff()}: MLL20_25 with General Recreational Effort increasing by 2\% per year

\item \code{MLL25_25()}: StatusQuo with Minimum Legal Length for both stocks (25 inch for Red Snapper and 25 inch for Gag)

\item \code{MLL25_25_IncRecEff()}: MLL25_25 with General Recreational Effort increasing by 2\% per year

}}
