% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPs.R
\name{Write_MPs}
\alias{Write_MPs}
\alias{Write_MPs_effort_creep}
\alias{Source_MPs}
\title{Write MP functions}
\usage{
Write_MPs(MPList, Rec_Reduction, dir = "MP_functions")

Write_MPs_effort_creep(
  MPList,
  Rec_Reduction,
  dir = "MP_functions_effort_creep"
)

Source_MPs(dir = "MP_functions")
}
\arguments{
\item{MPList}{A character string with names of the MPs to create}

\item{Rec_Reduction}{A numeric vector with effort reductions for General
Recreational fleet}

\item{dir}{Directory to write the MPs to}
}
\value{
A list of MPs names

A list of MPs names
}
\description{
Write MP functions
}
\section{Functions}{
\itemize{
\item \code{Write_MPs_effort_creep()}: Source the MPs with a 2\% effort creep

\item \code{Source_MPs()}: Source the MPs

}}
